pyproject.toml
[tool.poetry]
name = "event-strawberry"
version = "0.1.0"
description = "Event app backend API using django and strawberry-graphql"
authors = ["Wellington Zenon"]

[tool.poetry.dependencies]
python = "^3.8"
strawberry-graphql = "^0.71.0"
django-autoslug = "^1.9.8"
django-cors-headers = "^3.7.0"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
django-extensions = "^3.1.3"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
